backend/requirements.txt
finnhub-python
fastapi[standard]
uvicorn[standard]
python-dotenv
pydantic
requests